+++ src/issx/clients/exceptions.py
@@ -1,0 +1,10 @@
+class EntityDoesNotExistError(Exception):
+    pass
+
+
+class IssueDoesNotExistError(EntityDoesNotExistError):
+    pass
+
+
+class ProjectDoesNotExistError(EntityDoesNotExistError):
+    pass